Navigation Bug behoben

+++ Einkaufslisten_Template10/Models/Enum/TargetView.cs
@@ -6,8 +6,7 @@ using System.Threading.Tasks;
 
 namespace Einkaufslisten_Template10.Models.Enum
 {
-    public enum TargetView : int { EINKAUFEN = 0, LISTE = 1
-
-    }
-    
+    public enum TargetView : Byte {
+        EINKAUFEN = 0, LISTE = 1       
+    }  
 }
